[LOOP-5295] Delivery Log
