+++ Loop/Views/Insulin Delivery Log/InsulinDeliveryLog.swift
@@ -11,497 +11,6 @@ import LoopKit
 import LoopKitUI
 import SwiftUI
 
-fileprivate enum FilterOptions: Hashable, CaseIterable {
-    case userInitiated
-    case all
-    
-    var localizedMenuTitle: String {
-        switch self {
-        case .userInitiated:
-            NSLocalizedString("Self-Initiated Events", comment: "")
-        case .all:
-            NSLocalizedString("All Events", comment: "")
-        }
-    }
-}
-
-@MainActor
-@Observable
-class InsulinDeliveryLogViewModel {
-    
-    struct DisplayData: Hashable {
-        let insulinDeliveryState: InsulinDeliveryOverview.State, insulinDeliveryStateUpdatedDate: Date, currentBasalRate: DatedQuantity, lastAutoBolus: DatedQuantity?, totalInsulinDelivered: LoopQuantity, events: [InsulinDeliveryLogEvent]
-    }
-    
-    enum State: Hashable {
-        enum FetchError {
-            case noBasalRateSchedule
-        }
-        
-        case loading
-        case fetched(DisplayData)
-        case refreshing(DisplayData)
-        case error(FetchError)
-    }
-    
-    let totalDeliveredFormatter: QuantityFormatter = {
-        let formatter = QuantityFormatter(for: .internationalUnit)
-        
-        formatter.numberFormatter.maximumFractionDigits = 1
-        
-        return formatter
-    }()
-    
-    private let loopDataManager: LoopDataManager
-    private let pumpManager: PumpManager?
-    
-    private(set) var state: State
-    
-    fileprivate var selectedFilterOption: FilterOptions = .all
-    
-    var logEventDisplays: [LogEventDisplay] {
-        var displayEvents: [LogEventDisplay] = []
-        
-        switch state {
-        case .fetched(let data), .refreshing(let data):
-            data.events.filter {
-                switch selectedFilterOption {
-                case .userInitiated:
-                    switch $0.type {
-                    case .automation,
-                            .preset,
-                            .pumpEvent(.basal(.manualTempBasal, rate: _), _),
-                            .pumpEvent(.insulin, _),
-                            .pumpEvent(.bolus(.correction, _, _), _),
-                            .pumpEvent(.bolus(.meal, _, _), _):
-                        return true
-                    default:
-                        return false
-                    }
-                case .all:
-                    return true
-                }
-            }.filter {
-                $0.date >= Date().addingTimeInterval(.days(-1))
-            }.segmentItemsByHour().forEach { events in
-                displayEvents.append(.title(id: UUID(), "\(events.start.formatted(date: .omitted, time: .shortened)) - \(events.end.formatted(date: .omitted, time: .shortened))"))
-                events.events.forEach { event in
-                    displayEvents.append(.event(event))
-                }
-            }
-        case .loading, .error:
-            break
-        }
-        
-        return displayEvents
-    }
-    
-    var eventCount: Int {
-        logEventDisplays.filter { display in
-            switch display {
-            case .event:
-                return true
-            case .title:
-                return false
-            }
-        }.count
-    }
-    
-    private var doseStoreObserver: Any? {
-        willSet {
-            if let observer = doseStoreObserver {
-                NotificationCenter.default.removeObserver(observer)
-            }
-        }
-    }
-    
-    private var doseStore: DoseStore! {
-        didSet {
-            if let doseStore = doseStore {
-                doseStoreObserver = NotificationCenter.default.addObserver(forName: nil, object: doseStore, queue: OperationQueue.main, using: { [weak self] note in
-
-                    switch note.name {
-                    case DoseStore.valuesDidChange:
-                        Task { @MainActor in
-                            await self?.fetchData()
-                        }
-                    default:
-                        break
-                    }
-                })
-            } else {
-                doseStoreObserver = nil
-            }
-        }
-    }
-    
-    init(
-        loopDataManager: LoopDataManager,
-        pumpManager: PumpManager?,
-        initialState: State = .loading
-    ) {
-        self.loopDataManager = loopDataManager
-        self.pumpManager = pumpManager
-        self.state = initialState
-        
-        self.doseStore = (loopDataManager.doseStore as? DoseStore)
-        
-        Task {
-            await fetchData()
-        }
-    }
-
-    func fetchData() async {
-        if case let .fetched(data) = state {
-            state = .refreshing(data)
-        }
-        
-        var insulinDeliveryState: InsulinDeliveryOverview.State
-        var totalInsulinDelivered: LoopQuantity
-        var currentBasalRate: DatedQuantity
-        var lastAutoBolus: DatedQuantity?
-        var events = [InsulinDeliveryLogEvent]()
-        
-        let fetchedDate = Date()
-        let startDate = fetchedDate.addingTimeInterval(.days(-1))
-
-        // Status State
-        var insulinSuspended = false
-        if case .suspended = pumpManager?.status.basalDeliveryState {
-            insulinSuspended = true
-        }
-        
-        let automationEnabled = loopDataManager.automaticDosingStatus.automaticDosingEnabled
-        let automatedTreatmentState = pumpManager?.pumpManagerDelegate?.automatedTreatmentState ?? .neutralNoOverride
-
-        if insulinSuspended {
-            insulinDeliveryState = .error(status: .suspended)
-        } else if automationEnabled {
-            let basalStatus: InsulinDeliveryOverview.State.AutomatedBasalStatus
-            switch automatedTreatmentState {
-            case .neutralNoOverride, .neutralOverride:
-                basalStatus = .scheduled
-            case .increasedInsulin:
-                basalStatus = .moreThanScheduled
-            case .decreasedInsulin, .minimumDelivery:
-                basalStatus = .lessThanScheduled
-            }
-            
-            insulinDeliveryState = .automationOn(basalStatus: basalStatus, preset: loopDataManager.temporaryPresetsManager.activePreset)
-        } else {
-            insulinDeliveryState = .automationOff
-        }
-        
-        // Current Basal Rate
-        guard let basalRateSchedule = loopDataManager.temporaryPresetsManager.basalRateScheduleApplyingOverrideHistory ?? loopDataManager.settings.basalRateSchedule else {
-            state = .error(.noBasalRateSchedule)
-            return
-        }
-    
-        let currentValue = basalRateSchedule.scheduleSegment(at: startDate)
-        currentBasalRate = DatedQuantity(date: currentValue.startDate, quantity: LoopQuantity(unit: .internationalUnitsPerHour, doubleValue: currentValue.value))
-
-        // Basal and Bolus
-        let doses: [DoseEntry] = (try? await loopDataManager.doseStore.getNormalizedDoseEntries(start: startDate, end: nil)) ?? []
-
-        // Last Auto Bolus
-        if let lastAutoBolusDose = doses.filter({ $0.automatic == true }).last {
-            lastAutoBolus = DatedQuantity(date: lastAutoBolusDose.startDate, quantity: LoopQuantity(unit: .internationalUnit, doubleValue: lastAutoBolusDose.deliveredUnits ?? lastAutoBolusDose.value))
-        }
-        
-        // Total Insulin Delivered
-        totalInsulinDelivered = await LoopQuantity(unit: .internationalUnit, doubleValue: loopDataManager.totalDeliveredToday()?.value ?? 0)
-        
-        // Insulin Events
-        for dose in doses {
-            let automationEnabledDuringDose = loopDataManager.automationHistory.automationEnabled(at: dose.startDate) ?? loopDataManager.automaticDosingStatus.automaticDosingEnabled
-            
-            var decision: StoredDosingDecision?
-            if let decisionId = dose.decisionId {
-                decision = try? await loopDataManager.dosingDecisionStore.findDosingDecisionsById(decisionId)
-            }
-            
-            switch dose.type {
-            case .basal, .tempBasal:
-                if dose.type == .tempBasal && dose.automatic == false {
-                    events.append(
-                        InsulinDeliveryLogEvent(
-                            id: dose.syncIdentifier ?? UUID().uuidString,
-                            type: .pumpEvent(
-                                .basal(
-                                    .manualTempBasal(endDate: dose.endDate),
-                                    rate: LoopQuantity(
-                                        unit: .internationalUnitsPerHour,
-                                        doubleValue: dose.unitsPerHour
-                                    )
-                                ),
-                                dose
-                            ),
-                            date: dose.startDate
-                        )
-                    )
-                } else if automationEnabledDuringDose {
-                    if let decision {
-                        if decision.scheduleOverride != nil {
-                            events.append(
-                                InsulinDeliveryLogEvent(
-                                    id: dose.syncIdentifier ?? UUID().uuidString,
-                                    type: .pumpEvent(
-                                        .basal(
-                                            .automatedPresetBasal,
-                                            rate: LoopQuantity(
-                                                unit: .internationalUnitsPerHour,
-                                                doubleValue: dose.unitsPerHour
-                                            )
-                                        ),
-                                        dose
-                                    ),
-                                    date: dose.startDate
-                                )
-                            )
-                        } else {
-                            if let direction = decision.automaticDoseRecommendation?.direction {
-                                switch direction {
-                                case .decrease:
-                                    events.append(
-                                        InsulinDeliveryLogEvent(
-                                            id: dose.syncIdentifier ?? UUID().uuidString,
-                                            type: .pumpEvent(
-                                                .basal(
-                                                    .automationOn(basalStatus: .lessThanScheduled),
-                                                    rate: LoopQuantity(
-                                                        unit: .internationalUnitsPerHour,
-                                                        doubleValue: dose.unitsPerHour
-                                                    )
-                                                ),
-                                                dose
-                                            ),
-                                            date: dose.startDate
-                                        )
-                                    )
-                                case .neutral:
-                                    events.append(
-                                        InsulinDeliveryLogEvent(
-                                            id: dose.syncIdentifier ?? UUID().uuidString,
-                                            type: .pumpEvent(
-                                                .basal(
-                                                    .automationOn(basalStatus: .scheduled),
-                                                    rate: LoopQuantity(
-                                                        unit: .internationalUnitsPerHour,
-                                                        doubleValue: dose.unitsPerHour
-                                                    )
-                                                ),
-                                                dose
-                                            ),
-                                            date: dose.startDate
-                                        )
-                                    )
-                                case .increase:
-                                    events.append(
-                                        InsulinDeliveryLogEvent(
-                                            id: dose.syncIdentifier ?? UUID().uuidString,
-                                            type: .pumpEvent(
-                                                .basal(
-                                                    .automationOn(basalStatus: .moreThanScheduled),
-                                                    rate: LoopQuantity(
-                                                        unit: .internationalUnitsPerHour,
-                                                        doubleValue: dose.unitsPerHour
-                                                    )
-                                                ),
-                                                dose
-                                            ),
-                                            date: dose.startDate
-                                        )
-                                    )
-                                }
-                            } else {
-                                fatalError("No `decision.automaticDoseRecommendation`")
-                            }
-                        }
-                    } else if let scheduledBasalRate = dose.scheduledBasalRate, scheduledBasalRate.doubleValue(for: .internationalUnitsPerHour) == dose.value {
-                        events.append(
-                            InsulinDeliveryLogEvent(
-                                id: dose.syncIdentifier ?? UUID().uuidString,
-                                type: .pumpEvent(
-                                    .basal(
-                                        .automationOn(basalStatus: .scheduled),
-                                        rate: LoopQuantity(
-                                            unit: .internationalUnitsPerHour,
-                                            doubleValue: dose.unitsPerHour
-                                        )
-                                    ),
-                                    dose
-                                ),
-                                date: dose.startDate
-                            )
-                        )
-                    } else {
-                        fatalError("No `decision` or `scheduledBasalRate`")
-                    }
-                } else {
-                    events.append(
-                        InsulinDeliveryLogEvent(
-                            id: dose.syncIdentifier ?? UUID().uuidString,
-                            type: .pumpEvent(
-                                .basal(
-                                    .automationOff,
-                                    rate: LoopQuantity(
-                                        unit: .internationalUnitsPerHour,
-                                        doubleValue: dose.unitsPerHour
-                                    )
-                                ),
-                                dose
-                            ),
-                            date: dose.startDate
-                        )
-                    )
-                }
-            case .bolus:
-                if dose.automatic == true {
-                    events.append(
-                        InsulinDeliveryLogEvent(
-                            id: dose.syncIdentifier ?? UUID().uuidString,
-                            type: .pumpEvent(
-                                .bolus(
-                                    .automated,
-                                    programmedAmount: LoopQuantity(
-                                        unit: .internationalUnit,
-                                        doubleValue: dose.programmedUnits
-                                    ),
-                                    deliveryAmount: LoopQuantity(
-                                        unit: .internationalUnit,
-                                        doubleValue: dose.deliveredUnits ?? dose.programmedUnits
-                                    )
-                                ),
-                                dose
-                            ),
-                            date: dose.startDate
-                        )
-                    )
-                } else {
-                    if let recommendedUnits = decision?.manualBolusRecommendation?.recommendation.amount {
-                        if let carbEntry = decision?.carbEntry {
-                            events.append(
-                                InsulinDeliveryLogEvent(
-                                    id: decision?.syncIdentifier.uuidString ?? UUID().uuidString,
-                                    type: .pumpEvent(
-                                        .bolus(
-                                            .meal(
-                                                recommendedAmount: LoopQuantity(
-                                                    unit: .internationalUnit,
-                                                    doubleValue: recommendedUnits
-                                                ),
-                                                carbAmount: LoopQuantity(
-                                                    unit: .gram,
-                                                    doubleValue: carbEntry.amount
-                                                ),
-                                                emoji: carbEntry.foodType ?? ""
-                                            ),
-                                            programmedAmount: LoopQuantity(
-                                                unit: .internationalUnit,
-                                                doubleValue: decision?.manualBolusRequested ?? 0
-                                            ),
-                                            deliveryAmount: LoopQuantity(
-                                                unit: .internationalUnit,
-                                                doubleValue: dose.deliveredUnits ?? dose.programmedUnits
-                                            )
-                                        ),
-                                        dose
-                                    ),
-                                    date: dose.startDate
-                                )
-                            )
-                        } else {
-                            events.append(
-                                InsulinDeliveryLogEvent(
-                                    id: decision?.syncIdentifier.uuidString ?? UUID().uuidString,
-                                    type: .pumpEvent(
-                                        .bolus(
-                                            .correction(
-                                                recommendedAmount: LoopQuantity(
-                                                    unit: .internationalUnit,
-                                                    doubleValue: recommendedUnits
-                                                )
-                                            ),
-                                            programmedAmount: LoopQuantity(
-                                                unit: .internationalUnit,
-                                                doubleValue: decision?.manualBolusRequested ?? 0
-                                            ),
-                                            deliveryAmount: LoopQuantity(
-                                                unit: .internationalUnit,
-                                                doubleValue: dose.deliveredUnits ?? dose.programmedUnits
-                                            )
-                                        ),
-                                        dose
-                                    ),
-                                    date: dose.startDate
-                                )
-                            )
-                        }
-                    } else {
-                        events.append(
-                            InsulinDeliveryLogEvent(
-                                id: dose.syncIdentifier ?? UUID().uuidString,
-                                type: .pumpEvent(
-                                    .bolus(
-                                        .correction(recommendedAmount: nil),
-                                        programmedAmount: nil,
-                                        deliveryAmount: LoopQuantity(
-                                            unit: .internationalUnit,
-                                            doubleValue: dose.deliveredUnits ?? dose.programmedUnits
-                                        )
-                                    ),
-                                    dose
-                                ),
-                                date: dose.startDate
-                            )
-                        )
-                    }
-                }
-            case .resume:
-                break
-            case .suspend:
-                events.append(InsulinDeliveryLogEvent(id: dose.syncIdentifier ?? UUID().uuidString, type: .pumpEvent(.insulin(.suspended), dose), date: dose.startDate))
-                
-                if !dose.isMutable || dose.endDate <= fetchedDate {
-                    events.append(InsulinDeliveryLogEvent(id: dose.syncIdentifier ?? UUID().uuidString, type: .pumpEvent(.insulin(.resumed), dose), date: dose.endDate))
-                }
-            }
-        }
-        
-        // Automation
-        loopDataManager.automationHistory.forEach { event in
-            if event.enabled {
-                events.append(InsulinDeliveryLogEvent(id: String(event.hashValue), type: .automation(.on), date: event.startDate))
-            } else {
-                events.append(InsulinDeliveryLogEvent(id: String(event.hashValue), type: .automation(.off(endDate: nil)), date: event.startDate))
-            }
-        }
-        
-        // Preset
-        loopDataManager.temporaryPresetsManager.presetHistory.recentEvents.filter({ $0.override.actualEndDate >= startDate }).forEach { event in
-            if let preset = loopDataManager.temporaryPresetsManager.selectablePresets.first(where: { $0.id == event.override.presetId }) {
-                events.append(InsulinDeliveryLogEvent(id: String(event.hashValue), type: .preset(.enabled, icon: preset.icon, name: preset.name), date: event.override.startDate))
-                
-                if event.override.hasFinished() {
-                    events.append(InsulinDeliveryLogEvent(id: String(event.hashValue), type: .preset(.disabled, icon: preset.icon, name: preset.name), date: event.override.actualEndDate))
-                }
-            }
-        }
-        
-        state = .fetched(.init(insulinDeliveryState: insulinDeliveryState, insulinDeliveryStateUpdatedDate: fetchedDate, currentBasalRate: currentBasalRate, lastAutoBolus: lastAutoBolus, totalInsulinDelivered: totalInsulinDelivered, events: events))
-    }
-}
-
-enum LogEventDisplay: Hashable, Identifiable {
-    case title(id: UUID, String)
-    case event(InsulinDeliveryLogEvent)
-    
-    var id: Int {
-        hashValue
-    }
-}
-
 struct InsulinDeliveryLog: View {
     
     @State private var viewModel: InsulinDeliveryLogViewModel
@@ -514,7 +23,6 @@ struct InsulinDeliveryLog: View {
         self.onTapGesture = onTapGesture
     }
     
-    @ViewBuilder
     private func totalInsulinDeliveredLabel(from total: LoopQuantity) -> some View {
         LabeledContent {
             Text(viewModel.totalDeliveredFormatter.string(from: total) ?? "Unknown")
@@ -538,7 +46,7 @@ struct InsulinDeliveryLog: View {
             }
             
             Picker("Filter", selection: $viewModel.selectedFilterOption) {
-                ForEach(FilterOptions.allCases, id: \.self) { option in
+                ForEach(InsulinDeliveryLogViewModel.FilterOptions.allCases, id: \.self) { option in
                     Text(option.localizedMenuTitle)
                         .tag(option)
                 }
